Updated back-end
Updated back-end code to support 2 types of contact form. One for logged out users and one for logged in users. The new form is the one for logged out users or users who are yet to sign up, this form includes the email and phone and name fields that are automatically attached with the logged in contact form.