Normal text scales, as does the margin.
Now normal text on a webpage scales in size with screen size. Again using clamp().

Media queries used to set margin for the text. When the screen is too small for sidebar and main content, the margin is equal to each side. However, when the screen is larger enough for sidebar and content to be side-by-side, the margin is skewed to the right to give the appearance it is centred.