Sidebar and Main Content flex
The sidebar and the main content flex with sidebar getting 20% of the screen and main content getting the other 80%.

The links and images within the sidebar shrink according to the space they have to maintain a similar appearance.

In future the sidebar will only be default visible if screen size is above 800px, in the case where it is not, for example a mobile user, a button will allow the user to open the sidebar and interact with the content within.