Sidebar open and closes.
Now the navigation menu can be accessed by clicking on the 3 bar icon, which allows the user to then interact with the sidebar and this sidebar can be closed by pressing the x icon.

When a page loads the sidebar is automatically hidden and the 3 bar icon needs to be clicked before it opens.