Fixed issue with link in sign up page.
The link at the bottom of the sign up page which prompts the user to navigate to the sign in page if they already have an account, wrapped onto the second line on bigger screens as it out-grew the max-width.

Added clamp which works perfectly, but could just set the max-width to the max-value in the clamp.